webserver/requirements.txt
flask>=2.0.0
pyYAML >= 5.4.1
PyPDF2==1.26.0